Board, Inquiry, Umbrella, User 들의 id 값 타입 변경 ( int -> Long ), 대여(결제) 비밀번호 보관 방식 변경 ( Session & Cookie -> DB ( 비밀번호 회수를 위해서 ))
Spring Boot Mobile Application Project Start D+37
Spring Boot Start D+78

+++ src/main/java/com/example/usan/controller/InquiryController.java
@@ -46,13 +46,13 @@ public class InquiryController {
 
     }
     @GetMapping("/{id}")
-    public String findById(@PathVariable int id, Model model) {
+    public String findById(@PathVariable Long id, Model model) {
         model.addAttribute("inquiry", inquiryService.inquiryDetail(id));
         return "inquiry/detail";
     }
 
     @GetMapping("/{id}/updateForm")
-    public String updateForm(@PathVariable int id, Model model ){
+    public String updateForm(@PathVariable Long id, Model model ){
         model.addAttribute("inquiry", inquiryService.inquiryDetail(id));
         return "inquiry/updateForm";
     }
