+++ src/main/java/com/example/usan/repository/LikeRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface LikeRepository extends JpaRepository<Reply,Integer> {
+public interface LikeRepository extends JpaRepository<Reply,Long> {
 
     @Modifying
     @Query(value = "INSERT INTO boardlike(userId,boardId) VALUES(?1,?2)",nativeQuery = true)
-    int lSave(int userId, int boardId) ;
+    int lSave(Long userId, Long boardId) ;
 
 }
