+++ src/main/java/com/example/usan/service/BoardService.java
@@ -51,7 +51,7 @@ public class BoardService {
     // DB내의 Board들을 갖고와서 @PageableDefault 내에서 지정하는 size 만큼 짤라서 return
 
     @Transactional(readOnly = true)
-    public Board boardDetail(int id) {
+    public Board boardDetail(Long id) {
         return boardRepository.findById(id)
                 .orElseThrow(() -> {
                     return new IllegalArgumentException("글 상세보기 실패: 아이디를 찾을 수 없습니다.");
@@ -60,13 +60,13 @@ public class BoardService {
     }
 
     @Transactional
-    public void boardDelete(int id) {
+    public void boardDelete(Long id) {
         boardRepository.deleteById(id);
     }
     // 해당 id 값의 Board를 DB에서 삭제
 
     @Transactional
-    public void boardModify(int id, Board requestBoard) {
+    public void boardModify(Long id, Board requestBoard) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> {
                     return new IllegalArgumentException("글 상세보기 실패: 아이디를 찾을 수 없습니다.");
@@ -82,13 +82,13 @@ public class BoardService {
     // DB에 기존 Repository Save를 사용하지않고 따로 제작힌 mSave를 사용하여 DB에 저장
 
     @Transactional
-    public void replyDelete(int replyId) {
+    public void replyDelete(Long replyId) {
         replyRepository.deleteById(replyId);
     }
     // 해당 replyId값을 가진 Reply를 DB에서 삭제
 
     @Transactional
-    public void likeHit(int id, Board requestBoard) {
+    public void likeHit(Long id, Board requestBoard) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> {
                     return new IllegalArgumentException("글 상세보기 실패: 아이디를 찾을 수 없습니다.");
@@ -97,11 +97,11 @@ public class BoardService {
         // 좋아요를 눌렀을 때 기존 좋아요 수에 +1
     }
     @Transactional
-    public void likeHit(int id, LikeSaveRequestDto likeSaveRequestDto) { likeRepository.lSave(likeSaveRequestDto.getUserId(), likeSaveRequestDto.getBoardId()); }
+    public void likeHit(Long id, LikeSaveRequestDto likeSaveRequestDto) { likeRepository.lSave(likeSaveRequestDto.getUserId(), likeSaveRequestDto.getBoardId()); }
     // 좋아요를 누른 User와 눌려진 Board의 정보를 따로 저장
 
     @Transactional
-    public void viewCount(int id ) {
+    public void viewCount(Long id ) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> {
                     return new IllegalArgumentException("글 상세보기 실패: 아이디를 찾을 수 없습니다.");
